handle: fix panic when reading less data than the buffer can hold
Slice's copy_from_slice method expects the source's size to be equal to
that of the destination or it will panic. Such a condition can occur if
less data than what is available as space in the buffer is read.
With this patch we slice the destination slice to avoid the size
mismatch and, hence, the panic.