handle: correctly return data from Feature.get_from()
The get_from() method used to retrieve a feature report is lacking the
proper copy out into the result buffer. Further, the API design is
different from the Handle's read() method which behaves conceptually in
the same way, in that the report ID is not returned.
This change fixes both issues. The function signature is amended to
return the report ID along with the size of the data read.