handle: use an initialized Vec in read_from
By using `Vec::with_capacity` we reserve the space but its actual length is 0,
thus we pass along an empty slice to the read function which is always going to
fail because there's nowhere to write data to.

Use the `vec!` macro to initialize the vector to the correct size while zeroing
out the contents so we have somewhere for the read function to write the data
into.